create a own CAL product "level"
